move USER_KEY and JWT_KEY out of the user provider
